migrate all faer tests to a central location
