+++ crates/argmin-math/src/faer_tests.rs
@@ -1,5 +1,17 @@
 mod add;
+mod conj;
 mod div;
 mod dot;
 mod eye;
+mod inv;
+mod l1norm;
+mod l2norm;
+mod minmax;
+mod mul;
+mod random;
+mod signum;
+mod sub;
+mod transpose;
+mod zero;
+
 pub mod test_helper;
